♻️refactor/#10: ColumnService.readAll() 로직 일부를 Column으로 이동
