+++ BE/app/src/main/java/com/todo/app/domain/column/repository/CardRepository.java
@@ -1,6 +1,6 @@
 package com.todo.app.domain.column.repository;
 
-import com.todo.app.domain.column.entity.Card;
+import com.todo.app.domain.column.domain.Card;
 import java.util.List;
 
 public interface CardRepository {
