+++ BE/app/src/main/java/com/todo/app/domain/column/service/ColumnService.java
@@ -1,7 +1,7 @@
 package com.todo.app.domain.column.service;
 
-import com.todo.app.domain.column.entity.Card;
-import com.todo.app.domain.column.entity.Column;
+import com.todo.app.domain.column.domain.Card;
+import com.todo.app.domain.column.domain.Column;
 import com.todo.app.domain.column.repository.CardRepository;
 import com.todo.app.domain.column.repository.ColumnRepository;
 import java.util.ArrayList;
@@ -20,21 +20,14 @@ public class ColumnService {
         this.cardRepository = cardRepository;
     }
 
-    // TODO: 로직 개선 필요...
     public List<Column> readAll(Long memberId) {
         List<Column> columns = columnRepository.findAllBy(memberId);
         List<Card> cards = cardRepository.findAllBy(memberId);
 
         for (Column column : columns) {
-            List<Card> tmp = new ArrayList<>();
-
-            for (Card card : cards) {
-                if (Objects.equals(column.getId(), card.getColumnId())) {
-                    tmp.add(card);
-                }
-            }
-            column.setCards(tmp);
+            column.matchCards(cards);
         }
+        
         return columns;
     }
